Lazy load draft page sections
